rename folders and add index
